more changes for new db adapters

+++ db/migrate/20141205172800_add_details_to_currency.rb
@@ -1,6 +1,6 @@
 class AddDetailsToCurrency < ActiveRecord::Migration
   def change
-    add_column :currencies, :weight, :decimal
-    add_column :currencies, :collector_value, :decimal
+    add_column :currencies, :weight, :float
+    add_column :currencies, :collector_value, :float
   end
 end
